[ID-182] Login API issues (#188)
* in progress

* Setting def.yaml for Login Response

* in progress

* auth_type in Response for Login API

* setting the Changelog.md

* def.yaml Login Api

* Set ParamsResponse for the Login API

* resolve comment "account-preferences"

* fix the Response

* fix account/profile and accout/preferences

* fix response for the Login Api

* request the auth_type

* fix

* loading all the auth types

* set LoadAuthTypes interface

* fix the function name and set in core/auth/apis.go

* fixing comment

* load all auth types

* move loading authTypeFields from core to the web

* get the login auth type

* second for loop for identifing the auth types

* Remove empty line

* Fix the docs

* Update CHANGELOG.md

* Add auth_type_code field in accountAuthType - storage

* Return only the used auth type for the current login, not the whole list

* Fix the params response on the Login API when using illinois_oidc type. It was changed along with 155 by mistake.

* Add preferences param of the register user function

* Remove unused code

* Clean up the sign up flow - prepare the registration data in one place.

* Pass profile when registering a user

* Fix the birth year field

* Pass preferences when registering a user

* Update the docs

* set def.yaml. Add Profile and Preferences

* Login Request

* setting def.yaml for Profile and Preferences

* Setting the index

* set Preferences

* set Pforile fields

* Set Profil def.yaml and gen.types

* set def.yaml and gen.types for preferences

* profileDataFields

* set profile

* set preferences

* remove a return

* resolve comment. Remove json fields

* set def.yaml and gen.types

* set def.yaml

* remove

* remove the typos and  handle the profile and preferences with applyExternalAuthType()

* Fix empty line

* profileData is renamed to profile; added comments

* Reuse ProfileFields.yaml

* Fix API to work

* Fix the docs

+++ core/model/user.go
@@ -154,7 +154,7 @@ type Profile struct {
 	LastName  string
 	Email     string
 	Phone     string
-	BirthYear int8
+	BirthYear int16
 	Address   string
 	ZipCode   string
 	State     string
